+++ admin/controllers/ProductReviewsController.php
@@ -289,7 +289,7 @@ class ProductReviewsController extends ListingBaseController
             LibHelper::exitWithError(current($frm->getValidationErrors()), true);
         }
 
-        $data = SelProdReview::getAttributesById($recordId, ['spreview_id', 'spreview_status', 'spreview_lang_id']);
+        $data = SelProdReview::getAttributesById($recordId, ['spreview_id', 'spreview_status', 'spreview_lang_id', 'spreview_seller_user_id']);
         if (false == $data) {
             LibHelper::exitWithError($this->str_invalid_request, true);
         }
@@ -298,6 +298,10 @@ class ProductReviewsController extends ListingBaseController
         if (!$record->save()) {
             LibHelper::exitWithError($record->getError(), true);
         }
+
+        SelProdRating::updateSellerRating($data['spreview_seller_user_id']);
+        SelProdReview::updateSellerTotalReviews($data['spreview_seller_user_id']);
+
         $emailNotificationObj = new EmailHandler();
         $emailNotificationObj->sendBuyerReviewStatusUpdatedNotification($recordId, $data['spreview_lang_id']);