+++ admin/views/seller-package-plans/search.php
@@ -13,10 +13,23 @@ foreach ($arrListing as $sn => $row) {
     if ($row['spplan_active'] != applicationConstants::ACTIVE) {
         $tr->setAttribute("class", " nodrag nodrop");
     }
+
+    if ($row['spplan_active'] == applicationConstants::ACTIVE) {
+        $tr->setAttribute("id", $row['spplan_id']);
+    }
+
     foreach ($fields as $key => $val) {
         $tdAttr = ('action' == $key) ? ['class' => 'align-right'] : (('select_all' == $key) ? ['class' => 'col-check'] : []);
         $td = $tr->appendElement('td', $tdAttr);
         switch ($key) {
+            case 'dragdrop':
+                $div = $td->appendElement('div', ['class' => 'handleJs']);
+                $div->appendElement('plaintext', $tdAttr, '<svg class="svg" width="18" height="18">
+                                                            <use
+                                                                xlink:href="' . CONF_WEBROOT_URL . 'images/retina/sprite-actions.svg#drag">
+                                                            </use>
+                                                        </svg>', true);
+                break;
             case 'select_all':
                 $td->appendElement('plaintext', $tdAttr, '<label class="checkbox"><input class="selectItemJs" type="checkbox" name="record_ids[]" value=' . $row['spplan_id'] . '><i class="input-helper"></i></label>', true);
                 break;
@@ -64,7 +77,7 @@ foreach ($arrListing as $sn => $row) {
     $serialNo++;
 }
 
-include (CONF_THEME_PATH . '_partial/listing/no-record-found.php');
+include(CONF_THEME_PATH . '_partial/listing/no-record-found.php');
 
 if ($printData) {
     echo $tbody->getHtml();