+++ application/utilities/LibHelper.php
@@ -403,7 +403,8 @@ class LibHelper extends FatUtility
         // url check
         $parts = parse_url($url);
         if ($parts === false) {
-            throw new Exception('Unable to parse URL');
+            SystemLog::system('Unable to parse URL', 'Send Add To Cart Async Request', SystemLog::TYPE_ERROR);
+            return;
         }
 
         $host = $parts['host'] ?? null;
@@ -413,12 +414,14 @@ class LibHelper extends FatUtility
         parse_str($query, $queryParts);
 
         if ($host === null) {
-            throw new Exception('Unknown host');
+            SystemLog::system('Unknown Host', 'Send Add To Cart Async Request', SystemLog::TYPE_ERROR);
+            return;
         }
 
         $connection = fsockopen($host, $port, $errno, $errstr, 30);
         if ($connection === false) {
-            throw new Exception('Unable to connect to ' . $host);
+            SystemLog::system('Unable To Connect ' . $host, 'Send Add To Cart Async Request', SystemLog::TYPE_ERROR);
+            return;
         }
 
         $method = strtoupper($method);