+++ application/views/common-js-src/zdeveloper-common.js
@@ -1274,6 +1274,8 @@ $(function () {
     });
 
     $(".cc-cookie-accept-js").on('click', function () {
+        $("#cookieInfoBox").hide();
+        return;
         var data = {
             statistical_cookies: 1,
             personalise_cookies: 1,
@@ -1306,7 +1308,7 @@ $(function () {
         };
         updateUserCookies(data);
     };
-    updateUserCookies = function (data) {
+    updateUserCookies = function (data) {        
         fcom.ajax(
             fcom.makeUrl("Custom", "updateUserCookies"),
             data,
@@ -1314,6 +1316,7 @@ $(function () {
                 var ans = $.parseJSON(rsp);
                 if (ans.status == 0) {
                     fcom.displayErrorMessage(ans.msg);
+                    $("#cookieInfoBox").show();
                 } else {
                     $("#cookieInfoBox").hide("slow");
                     $("#cookieInfoBox").remove();